requirements.txt
pyautogen[gemini,retrievechat,lmm]